Serviços de Prefeitura alinhados.

+++ src/main/java/com/ufcg/si1/model/prefeitura/Caos.java
@@ -3,7 +3,7 @@ package com.ufcg.si1.model.prefeitura;
 public class Caos implements SituacaoDaPrefeitura {
 
 	@Override
-	public SituacaoGeralDasQueixas getSituacaoDasQueixas(float porcentagemQueixasAbertas) {
+	public SituacaoGeralDasQueixas getSituacaoDasQueixas(double porcentagemQueixasAbertas) {
 		if (porcentagemQueixasAbertas > 5) {
 			return SituacaoGeralDasQueixas.RUIM;
 		} else if (porcentagemQueixasAbertas > 2) {
