+++ src/main/java/com/ufcg/si1/model/prefeitura/Normal.java
@@ -3,7 +3,7 @@ package com.ufcg.si1.model.prefeitura;
 public class Normal implements SituacaoDaPrefeitura {
 
 	@Override
-	public SituacaoGeralDasQueixas getSituacaoDasQueixas(float porcentagemQueixasAbertas) {
+	public SituacaoGeralDasQueixas getSituacaoDasQueixas(double porcentagemQueixasAbertas) {
 		if (porcentagemQueixasAbertas > 20) {
 			return SituacaoGeralDasQueixas.RUIM;
 		} else if (porcentagemQueixasAbertas > 10) {
