+++ src/main/java/com/ufcg/si1/model/prefeitura/SituacaoDaPrefeitura.java
@@ -2,6 +2,6 @@ package com.ufcg.si1.model.prefeitura;
 
 public interface SituacaoDaPrefeitura {
 
-	public SituacaoGeralDasQueixas getSituacaoDasQueixas(float porcentagemQueixasAbertas);
+	public SituacaoGeralDasQueixas getSituacaoDasQueixas(double porcentagemQueixasAbertas);
 
 }
